Fix lifecycle email worker boundary
