+++ apps/server/workers/src/processors/api/queues/lifecycle-email/lifecycle-email.processor.ts
@@ -1,4 +1,3 @@
-import { LifecycleEmailService } from '@api/services/lifecycle-emails/lifecycle-email.service';
 import {
   LIFECYCLE_EMAIL_QUEUE,
   type LifecycleEmailJobData,
@@ -10,6 +9,7 @@ import {
   type ProcessorCircuitBreaker,
 } from '@libs/utils/circuit-breaker/circuit-breaker.util';
 import { Processor, WorkerHost } from '@nestjs/bullmq';
+import { LifecycleEmailDeliveryService } from '@server/services/lifecycle-emails/lifecycle-email-delivery.service';
 import type { Job } from 'bullmq';
 
 @Processor(LIFECYCLE_EMAIL_QUEUE)
@@ -17,7 +17,7 @@ export class LifecycleEmailProcessor extends WorkerHost {
   private readonly circuitBreaker: ProcessorCircuitBreaker;
 
   constructor(
-    private readonly lifecycleEmailService: LifecycleEmailService,
+    private readonly lifecycleEmailDeliveryService: LifecycleEmailDeliveryService,
     private readonly logger: LoggerService,
   ) {
     super();
@@ -31,7 +31,7 @@ export class LifecycleEmailProcessor extends WorkerHost {
     try {
       await this.circuitBreaker.execute(async () => {
         await job.updateProgress(10);
-        await this.lifecycleEmailService.sendLifecycleEmail(job.data);
+        await this.lifecycleEmailDeliveryService.sendLifecycleEmail(job.data);
         await job.updateProgress(100);
       });
     } catch (error: unknown) {
